src/sheet_music_analyzer.py
"""
Sheet Music Analyzer - Uses AI/ML to extract musical information from PDF sheet music
"""
import os
import logging
from typing import Dict, List, Any
import fitz  # PyMuPDF
from PIL import Image
import numpy as np
import cv2

logger = logging.getLogger(__name__)


class SheetMusicAnalyzer:
    """Analyzes sheet music PDFs and extracts musical information"""
    
    def __init__(self):
        """Initialize the sheet music analyzer"""
        self.logger = logger
    
    def analyze(self, pdf_path: str) -> Dict[str, Any]:
        """
        Analyze a PDF sheet music file
        
        Args:
            pdf_path: Path to the PDF file
            
        Returns:
            Dictionary containing analysis results
        """
        try:
            self.logger.info(f"Starting analysis of {pdf_path}")
            
            # Extract images from PDF
            images = self._extract_images_from_pdf(pdf_path)
            
            # Perform OMR (Optical Music Recognition) on each page
            notes = []
            rhythms = []
            time_signature = None
            key_signature = None
            tempo = 120  # Default tempo
            
            for idx, image in enumerate(images):
                self.logger.info(f"Processing page {idx + 1}")
                page_analysis = self._analyze_page(image)
                
                notes.extend(page_analysis['notes'])
                rhythms.extend(page_analysis['rhythms'])
                
                # Get metadata from first page
                if idx == 0:
                    time_signature = page_analysis.get('time_signature', '4/4')
                    key_signature = page_analysis.get('key_signature', 'C')
                    tempo = page_analysis.get('tempo', 120)
            
            analysis = {
                'notes': notes,
                'rhythms': rhythms,
                'time_signature': time_signature,
                'key_signature': key_signature,
                'tempo': tempo,
                'num_pages': len(images),
                'total_measures': len(rhythms) // 4 if rhythms else 0  # Approximate
            }
            
            self.logger.info(f"Analysis complete: {len(notes)} notes found")
            return analysis
            
        except Exception as e:
            self.logger.error(f"Error analyzing sheet music: {str(e)}")
            raise
    
    def _extract_images_from_pdf(self, pdf_path: str) -> List[np.ndarray]:
        """Extract images from PDF pages"""
        images = []
        
        try:
            # Open PDF
            pdf_document = fitz.open(pdf_path)
            
            for page_num in range(len(pdf_document)):
                page = pdf_document[page_num]
                
                # Render page to image at high resolution
                mat = fitz.Matrix(2.0, 2.0)  # 2x zoom for better quality
                pix = page.get_pixmap(matrix=mat)
                
                # Convert to numpy array
                img = Image.frombytes("RGB", [pix.width, pix.height], pix.samples)
                img_array = np.array(img)
                
                images.append(img_array)
            
            pdf_document.close()
            return images
            
        except Exception as e:
            self.logger.error(f"Error extracting images from PDF: {str(e)}")
            raise
    
    def _analyze_page(self, image: np.ndarray) -> Dict[str, Any]:
        """
        Analyze a single page of sheet music
        This is a simplified implementation - a full production system would use
        a trained ML model for accurate OMR
        """
        # Convert to grayscale
        gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)
        
        # Apply thresholding
        _, binary = cv2.threshold(gray, 127, 255, cv2.THRESH_BINARY_INV)
        
        # Detect staff lines
        staff_lines = self._detect_staff_lines(binary)
        
        # For this implementation, we'll create a basic structure
        # In production, this would use a trained ML model (like Audiveris, OMR models)
        page_analysis = {
            'notes': self._extract_notes_basic(binary, staff_lines),
            'rhythms': self._extract_rhythms_basic(binary, staff_lines),
            'time_signature': '4/4',
            'key_signature': 'C',
            'tempo': 120,
            'staff_lines': len(staff_lines)
        }
        
        return page_analysis
    
    def _detect_staff_lines(self, binary_image: np.ndarray) -> List[int]:
        """Detect horizontal staff lines in the image"""
        # Use horizontal projection to find staff lines
        horizontal_projection = np.sum(binary_image, axis=1)
        
        # Find peaks (staff lines have high values)
        threshold = np.max(horizontal_projection) * 0.7
        staff_lines = []
        
        for i, value in enumerate(horizontal_projection):
            if value > threshold:
                # Avoid duplicates by checking distance from last line
                if not staff_lines or i - staff_lines[-1] > 10:
                    staff_lines.append(i)
        
        return staff_lines
    
    def _extract_notes_basic(self, binary_image: np.ndarray, staff_lines: List[int]) -> List[Dict]:
        """
        Basic note extraction
        In production, this would use ML model trained on sheet music
        """
        # This is a placeholder - would be replaced with actual OMR
        # For demonstration, create a sample note sequence
        notes = []
        
        # Generate a basic C major scale as example
        note_names = ['C4', 'D4', 'E4', 'F4', 'G4', 'A4', 'B4', 'C5']
        for i, note in enumerate(note_names):
            notes.append({
                'pitch': note,
                'start_time': i * 0.5,
                'duration': 0.5,
                'position': i
            })
        
        return notes
    
    def _extract_rhythms_basic(self, binary_image: np.ndarray, staff_lines: List[int]) -> List[Dict]:
        """
        Basic rhythm extraction
        In production, this would use ML model trained on sheet music
        """
        # Placeholder rhythm data
        rhythms = []
        
        # Generate basic quarter notes
        for i in range(8):
            rhythms.append({
                'type': 'quarter',
                'duration': 1.0,
                'position': i
            })
        
        return rhythms
