requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==22.0.0

# PDF and Sheet Music Processing
pdf2image==1.16.3
Pillow==10.3.0
PyMuPDF==1.23.8

# AI/ML for OMR (Optical Music Recognition)
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0

# Audio Processing and Analysis
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4
noisereduce==3.0.0
basic-pitch==0.2.5

# Music Analysis
music21==9.1.0
pretty_midi==0.2.10
mido==1.3.0

# Open Source LLM for text generation
transformers==4.48.0
torch==2.6.0
accelerate==0.25.0
sentencepiece==0.1.99

# Authentication & Security
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
email-validator==2.1.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0

# Note: Audiveris requires Java Runtime Environment (JRE) 11 or higher
# The Dockerfile installs OpenJDK 17 which is compatible
# See DEPLOYMENT.md for Audiveris installation instructions
